+++ src/lexers/javascript-lexer.js
@@ -67,6 +67,10 @@ export default class JavascriptLexer extends BaseLexer {
 
       parseCommentNode(keys, node, content)
 
+      if (node.kind === ts.SyntaxKind.TaggedTemplateExpression) {
+        entry = this.taggedTemplateExpressionExtractor.call(this, node)
+      }
+
       if (node.kind === ts.SyntaxKind.CallExpression) {
         entry = this.expressionExtractor.call(this, node)
       }
@@ -86,6 +90,30 @@ export default class JavascriptLexer extends BaseLexer {
     parseTree(sourceFile)
 
     return this.setNamespaces(keys)
+  }
+
+  taggedTemplateExpressionExtractor(node) {
+    const entry = {}
+
+    const { tag, template } = node
+
+    const isTranslationFunction =
+      (tag.text && this.functions.includes(tag.text)) ||
+      (tag.name && this.functions.includes(tag.name.text))
+
+    if (!isTranslationFunction) return null
+
+    if (template.kind === ts.SyntaxKind.NoSubstitutionTemplateLiteral) {
+      entry.key = template.text
+    } else if (template.kind === ts.SyntaxKind.TemplateExpression) {
+      this.emit(
+        'warning',
+        'A key that is a template string must not have any interpolations.'
+      )
+      return null
+    }
+
+    return entry
   }
 
   expressionExtractor(node) {
@@ -138,12 +166,12 @@ export default class JavascriptLexer extends BaseLexer {
         }
         entry.key = concatenatedString
       } else {
-        if (keyArgument.kind === ts.SyntaxKind.Identifier) {
-          this.emit(
-            'warning',
-            `Key is not a string literal: ${keyArgument.text}`
-          )
-        }
+        this.emit(
+          'warning',
+          keyArgument.kind === ts.SyntaxKind.Identifier
+            ? `Key is not a string literal: ${keyArgument.text}`
+            : 'Key is not a string literal'
+        )
         return null
       }
 
@@ -163,7 +191,14 @@ export default class JavascriptLexer extends BaseLexer {
         optionsArgument.kind === ts.SyntaxKind.ObjectLiteralExpression
       ) {
         for (const p of optionsArgument.properties) {
-          entry[p.name.text] = (p.initializer && p.initializer.text) || ''
+          if (p.kind === ts.SyntaxKind.SpreadAssignment) {
+            this.emit(
+              'warning',
+              `Options argument is a spread operator : ${p.expression.text}`
+            )
+          } else {
+            entry[p.name.text] = (p.initializer && p.initializer.text) || ''
+          }
         }
       }