+++ src/lexers/vue-lexer.js
@@ -15,8 +15,9 @@ export default class VueLexer extends BaseLexer {
     Lexer.on('warning', (warning) => this.emit('warning', warning))
     keys = keys.concat(Lexer.extract(content))
 
-    const compiledTemplate = require('vue-template-compiler').compile(content)
-      .render
+    const compiledTemplate = require('vue-template-compiler').compile(
+      content
+    ).render
     const Lexer2 = new JavascriptLexer({ functions: this.functions })
     Lexer2.on('warning', (warning) => this.emit('warning', warning))
     keys = keys.concat(Lexer2.extract(compiledTemplate))